golden-paths/h3-innovation/rag-application/skeleton/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.27.0
python-multipart>=0.0.27

# Azure SDKs
azure-identity>=1.19.0
azure-storage-blob==12.19.0
azure-search-documents==11.4.0
openai==1.12.0

# LangChain
langchain==0.3.21
langchain-openai>=0.2.14
langchain-community>=0.3.27

# Document processing
pypdf>=6.10.2
python-docx==1.1.0
beautifulsoup4==4.12.3
tiktoken==0.5.2

# Observability
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0

# Utilities
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv>=1.2.2
tenacity==8.2.3

# Testing
pytest>=9.0.3
pytest-asyncio==0.23.4
httpx==0.26.0
